changed bounded number converter contract to prevent malformed classes
